+++ src/main/java/com/jalaramrakhi/erpjr/repository/AccountTypeRepository.java
@@ -3,5 +3,5 @@ package com.jalaramrakhi.erpjr.repository;
 import com.jalaramrakhi.erpjr.entity.AccountType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccountTypeRepository extends JpaRepository<AccountType, Integer> {
+public interface AccountTypeRepository extends JpaRepository<AccountType, Long> {
 }
