+++ src/main/java/com/jalaramrakhi/erpjr/repository/BarcodeItemDetailRepository.java
@@ -3,5 +3,5 @@ package com.jalaramrakhi.erpjr.repository;
 import com.jalaramrakhi.erpjr.entity.BarcodeItemDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BarcodeItemDetailRepository extends JpaRepository<BarcodeItemDetails, Integer> {
+public interface BarcodeItemDetailRepository extends JpaRepository<BarcodeItemDetails, Long> {
 }
