+++ src/main/java/com/jalaramrakhi/erpjr/repository/InventoryItemRepository.java
@@ -3,5 +3,5 @@ package com.jalaramrakhi.erpjr.repository;
 import com.jalaramrakhi.erpjr.entity.InventoryItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InventoryItemRepository extends JpaRepository<InventoryItem, Integer> {
+public interface InventoryItemRepository extends JpaRepository<InventoryItem, Long> {
 }
