+++ src/main/java/com/jalaramrakhi/erpjr/repository/UnitRepository.java
@@ -3,5 +3,5 @@ package com.jalaramrakhi.erpjr.repository;
 import com.jalaramrakhi.erpjr.entity.Unit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UnitRepository extends JpaRepository<Unit, Integer> {
+public interface UnitRepository extends JpaRepository<Unit, Long> {
 }
