+++ src/main/java/com/jalaramrakhi/erpjr/repository/UserRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     @Query(value = "SELECT * FROM tbl_user WHERE user_name = ?1 AND company_id = ?2", nativeQuery = true)
-    User findUserByCompany(String username, Integer company);
+    User findUserByCompany(String username, Long company);
 }
