+++ src/main/java/com/jalaramrakhi/erpjr/service/AccountTypeService.java
@@ -8,8 +8,8 @@ import java.util.List;
 
 public interface AccountTypeService {
     public ResponseEntity<List<AccountType>> getAllAccountTypes();
-    public ResponseEntity<AccountType> getSingleAccountType(Integer id);
+    public ResponseEntity<AccountType> getSingleAccountType(Long id);
     public ResponseEntity<AccountType> addNewAccountType(AccountType accountType, HttpServletRequest request);
-    public ResponseEntity<AccountType> updateAccountType(Integer id, AccountType accountType);
-    public ResponseEntity<AccountType> deleteAccountType(Integer id);
+    public ResponseEntity<AccountType> updateAccountType(Long id, AccountType accountType);
+    public ResponseEntity<AccountType> deleteAccountType(Long id);
 }
